+++ src/Actions/ActionFunctions.hpp
@@ -121,8 +121,8 @@ namespace osc
     // attempts to reassign a component's socket connection (returns false and writes to `error` on failure)
     bool ActionReassignComponentSocket(UndoableModelStatePair&, OpenSim::ComponentPath const& componentAbsPath, std::string const& socketName, OpenSim::Object const& connectee, std::string& error);
 
-    // sets the model's isolation to the provided component (can be nullptr)
-    bool ActionSetModelIsolationTo(UndoableModelStatePair&, OpenSim::Component const*);
+    // sets the model's showing only to the provided component (can be nullptr)
+    bool ActionSetModelShowingOnlyTo(UndoableModelStatePair&, OpenSim::Component const*);
 
     // sets the model's scale factor
     bool ActionSetModelSceneScaleFactorTo(UndoableModelStatePair&, float);