+++ src/OpenSimBindings/OpenSimHelpers.cpp
@@ -97,7 +97,7 @@ namespace
         OpenSim::Component const& c,
         OpenSim::Component const* selected,
         OpenSim::Component const* hovered,
-        OpenSim::Component const* isolated)
+        OpenSim::Component const* showingOnly)
     {
         osc::SceneDecorationFlags rv = osc::SceneDecorationFlags_None;
 
@@ -111,9 +111,9 @@ namespace
             rv |= osc::SceneDecorationFlags_IsHovered;
         }
 
-        if (&c == isolated)
+        if (&c == showingOnly)
         {
-            rv |= osc::SceneDecorationFlags_IsIsolated;
+            rv |= osc::SceneDecorationFlags_IsShowingOnly;
         }
 
         OpenSim::Component const* ptr = osc::GetOwner(c);
@@ -127,9 +127,9 @@ namespace
             {
                 rv |= osc::SceneDecorationFlags_IsChildOfHovered;
             }
-            if (ptr == isolated)
+            if (ptr == showingOnly)
             {
-                rv |= osc::SceneDecorationFlags_IsChildOfIsolated;
+                rv |= osc::SceneDecorationFlags_IsChildOfShowingOnly;
             }
             ptr = osc::GetOwner(*ptr);
         }
@@ -278,7 +278,7 @@ namespace
                                   SimTK::State const& st,
                                   OpenSim::Component const* selected,
                                   OpenSim::Component const* hovered,
-                                  OpenSim::Component const* isolated,
+                                  OpenSim::Component const* showingOnly,
                                   float fixupScaleFactor,
                                   std::vector<osc::SceneDecoration>& out)
     {
@@ -293,7 +293,7 @@ namespace
             cylinderXform,
             glm::vec4{0.7f, 0.7f, 0.7f, 1.0f},
             p2p.getAbsolutePathString(),
-            ComputeFlags(p2p, selected, hovered, isolated)
+            ComputeFlags(p2p, selected, hovered, showingOnly)
         );
     }
 
@@ -302,7 +302,7 @@ namespace
                        SimTK::State const& st,
                        OpenSim::Component const* selected,
                        OpenSim::Component const* hovered,
-                       OpenSim::Component const* isolated,
+                       OpenSim::Component const* showingOnly,
                        float fixupScaleFactor,
                        std::vector<osc::SceneDecoration>& out)
     {
@@ -317,7 +317,7 @@ namespace
             xform,
             glm::vec4{1.0f, 0.0f, 0.0f, 1.0f},
             s.getAbsolutePathString(),
-            ComputeFlags(s, selected, hovered, isolated)
+            ComputeFlags(s, selected, hovered, showingOnly)
         );
     }
 
@@ -325,7 +325,7 @@ namespace
                                     SimTK::State const& st,
                                     OpenSim::Component const* selected,
                                     OpenSim::Component const* hovered,
-                                    OpenSim::Component const* isolated,
+                                    OpenSim::Component const* showingOnly,
                                     float fixupScaleFactor,
                                     std::vector<osc::SceneDecoration>& out)
     {
@@ -337,7 +337,7 @@ namespace
             t,
             glm::vec4{1.0f, 1.0f, 0.0f, 0.2f},
             j.getAbsolutePathString(),
-            ComputeFlags(j, selected, hovered, isolated)
+            ComputeFlags(j, selected, hovered, showingOnly)
         );
     }
 
@@ -347,7 +347,7 @@ namespace
                     float fixupScaleFactor,
                     OpenSim::Component const* selected,
                     OpenSim::Component const* hovered,
-                    OpenSim::Component const* isolated,
+                    OpenSim::Component const* showingOnly,
                     std::vector<osc::SceneDecoration>& out,
                     OpenSim::ModelDisplayHints const& mdh,
                     SimTK::Array_<SimTK::DecorativeGeometry>& geomList,
@@ -367,7 +367,7 @@ namespace
                 t,
                 glm::vec4{0.0f, 0.0f, 0.0f, 1.0f},
                 b.getAbsolutePathString(),
-                ComputeFlags(b, selected, hovered, isolated)
+                ComputeFlags(b, selected, hovered, showingOnly)
             );
         }
 
@@ -380,7 +380,7 @@ namespace
                                 SimTK::State const& st,
                                 OpenSim::Component const* selected,
                                 OpenSim::Component const* hovered,
-                                OpenSim::Component const* isolated,
+                                OpenSim::Component const* showingOnly,
                                 float fixupScaleFactor,
                                 OpenSim::ModelDisplayHints const&,
                                 std::vector<osc::SceneDecoration>& out)
@@ -400,7 +400,7 @@ namespace
         glm::vec4 const fiberColor = GetSconeStyleMuscleColor(muscle, st, opts.getMuscleColoringStyle());
         glm::vec4 const tendonColor = {204.0f/255.0f, 203.0f/255.0f, 200.0f/255.0f, 1.0f};
 
-        osc::SceneDecorationFlags flags = ComputeFlags(muscle, selected, hovered, isolated);
+        osc::SceneDecorationFlags flags = ComputeFlags(muscle, selected, hovered, showingOnly);
 
         osc::SceneDecoration fiberSpherePrototype =
         {
@@ -579,7 +579,7 @@ namespace
                                   SimTK::State const& st,
                                   OpenSim::Component const* selected,
                                   OpenSim::Component const* hovered,
-                                  OpenSim::Component const* isolated,
+                                  OpenSim::Component const* showingOnly,
                                   float fixupScaleFactor,
                                   OpenSim::Component const**,
                                   OpenSim::ModelDisplayHints const& mdh,
@@ -587,7 +587,7 @@ namespace
                                   osc::DecorativeGeometryHandler&,
                                   std::vector<osc::SceneDecoration>& out)
     {
-        osc::SceneDecorationFlags const flags = ComputeFlags(musc, selected, hovered, isolated);
+        osc::SceneDecorationFlags const flags = ComputeFlags(musc, selected, hovered, showingOnly);
         std::vector<glm::vec3> const pps = GetAllPathPoints(musc.getGeometryPath(), st);
         std::string const absPath = musc.getAbsolutePathString();
 
@@ -649,7 +649,7 @@ namespace
                             SimTK::State const& st,
                             OpenSim::Component const* selected,
                             OpenSim::Component const* hovered,
-                            OpenSim::Component const* isolated,
+                            OpenSim::Component const* showingOnly,
                             float fixupScaleFactor,
                             OpenSim::Component const** currentComponent,
                             OpenSim::ModelDisplayHints const& mdh,
@@ -667,13 +667,13 @@ namespace
                 switch (opts.getMuscleDecorationStyle())
                 {
                 case osc::MuscleDecorationStyle::Scone:
-                    HandleMuscleSconeStyle(opts, *musc, st, selected, hovered, isolated, fixupScaleFactor, mdh, out);
+                    HandleMuscleSconeStyle(opts, *musc, st, selected, hovered, showingOnly, fixupScaleFactor, mdh, out);
                     return;
                 case osc::MuscleDecorationStyle::Hidden:
                     return;  // just don't generate them
                 case osc::MuscleDecorationStyle::OpenSim:
                 default:
-                    HandleMuscleOpenSimStyle(opts, *musc, st, selected, hovered, isolated, fixupScaleFactor, currentComponent, mdh, geomList, producer, out);
+                    HandleMuscleOpenSimStyle(opts, *musc, st, selected, hovered, showingOnly, fixupScaleFactor, currentComponent, mdh, geomList, producer, out);
                     return;
                 }
             }
@@ -707,14 +707,14 @@ namespace
         void operator()(std::shared_ptr<osc::Mesh const> const& mesh, osc::Transform const& t, glm::vec4 const& color) override
         {
             std::string absPath = (*m_CurrentComponent) ? (*m_CurrentComponent)->getAbsolutePathString() : std::string{};
-            m_Out->emplace_back(mesh, t, color, std::move(absPath), ComputeFlags(**m_CurrentComponent, m_Selected, m_Hovered, m_Isolated));
+            m_Out->emplace_back(mesh, t, color, std::move(absPath), ComputeFlags(**m_CurrentComponent, m_Selected, m_Hovered, m_ShowingOnly));
         }
 
     private:
         osc::VirtualConstModelStatePair const* m_Msp;
         OpenSim::Component const* m_Selected = m_Msp->getSelected();
         OpenSim::Component const* m_Hovered = m_Msp->getHovered();
-        OpenSim::Component const* m_Isolated = m_Msp->getIsolated();
+        OpenSim::Component const* m_ShowingOnly = m_Msp->getShowingOnly();
         std::vector<osc::SceneDecoration>* m_Out;
         OpenSim::Component const** m_CurrentComponent;
     };
@@ -732,7 +732,7 @@ namespace
         SimTK::State const& state = msp.getState();
         OpenSim::Component const* selected = msp.getSelected();
         OpenSim::Component const* hovered = msp.getHovered();
-        OpenSim::Component const* isolated = msp.getIsolated();
+        OpenSim::Component const* showingOnly = msp.getShowingOnly();
         float fixupScaleFactor = msp.getFixupScaleFactor();
         OpenSim::ModelDisplayHints const& mdh = msp.getModel().getDisplayHints();
 
@@ -764,23 +764,23 @@ namespace
             // component decoration handling
             if (auto const* p2p = dynamic_cast<OpenSim::PointToPointSpring const*>(&c))
             {
-                HandlePointToPointSpring(*p2p, state, selected, hovered, isolated, fixupScaleFactor, out);
+                HandlePointToPointSpring(*p2p, state, selected, hovered, showingOnly, fixupScaleFactor, out);
             }
             else if (typeid(c) == typeid(OpenSim::Station))  // CARE: it's a typeid comparison because OpenSim::Marker inherits from OpenSim::Station
             {
-                HandleStation(static_cast<OpenSim::Station const&>(c), state, selected, hovered, isolated, fixupScaleFactor, out);
+                HandleStation(static_cast<OpenSim::Station const&>(c), state, selected, hovered, showingOnly, fixupScaleFactor, out);
             }
             else if (auto const* scapulo = dynamic_cast<OpenSim::ScapulothoracicJoint const*>(&c); scapulo && opts.getShouldShowScapulo())
             {
-                HandleScapulothoracicJoint(*scapulo, state, selected, hovered, isolated, fixupScaleFactor, out);
+                HandleScapulothoracicJoint(*scapulo, state, selected, hovered, showingOnly, fixupScaleFactor, out);
             }
             else if (auto const* body = dynamic_cast<OpenSim::Body const*>(&c))
             {
-                HandleBody(*body, state, fixupScaleFactor, selected, hovered, isolated, out, mdh, geomList, producer);
+                HandleBody(*body, state, fixupScaleFactor, selected, hovered, showingOnly, out, mdh, geomList, producer);
             }
             else if (auto const* gp = dynamic_cast<OpenSim::GeometryPath const*>(&c))
             {
-                HandleGeometryPath(opts, *gp, state, selected, hovered, isolated, fixupScaleFactor, &currentComponent, mdh, geomList, producer, out);
+                HandleGeometryPath(opts, *gp, state, selected, hovered, showingOnly, fixupScaleFactor, &currentComponent, mdh, geomList, producer, out);
             }
             else
             {