+++ src/OpenSimBindings/VirtualModelStatePair.hpp
@@ -21,7 +21,7 @@ namespace osc
 
         virtual void setHovered(OpenSim::Component const*) {}
 
-        virtual void setIsolated(OpenSim::Component const*) {}
+        virtual void setShowingOnly(OpenSim::Component const*) {}
 
         virtual void setFixupScaleFactor(float) {}
     };